Remove unnecessary config module and move Websocket
