+++ plugins/c8y_remote_access_plugin/src/main.rs
@@ -1,18 +1,16 @@
 use std::path::Path;
+use std::path::PathBuf;
 
-use async_tungstenite::tokio::ConnectStream;
-use config::supported_operation_path;
-use config::C8yUrl;
 use miette::Context;
 use miette::IntoDiagnostic;
-use rand::prelude::*;
+use tedge_config::C8yUrlSetting;
 use tedge_config::ConfigRepository;
+use tedge_config::ConfigSettingAccessor;
 use tedge_config::TEdgeConfig;
 use tedge_config::TEdgeConfigLocation;
 use tedge_config::TEdgeConfigRepository;
 use tedge_utils::file::create_file_with_user_group;
 use url::Url;
-use ws_stream_tungstenite::WsStream;
 
 use crate::auth::Jwt;
 use crate::input::Command;
@@ -20,7 +18,6 @@ use crate::input::RemoteAccessConnect;
 use crate::proxy::WebsocketSocketProxy;
 
 mod auth;
-mod config;
 mod csv;
 mod input;
 mod proxy;
@@ -68,8 +65,8 @@ fn remove_supported_operation(config_dir: &Path) -> miette::Result<()> {
 }
 
 async fn proxy(command: RemoteAccessConnect, config: TEdgeConfig) -> miette::Result<()> {
-    let C8yUrl(host) = C8yUrl::retrieve(&config)?;
-    let url = build_proxy_url(&host, command.key())?;
+    let host = config.query(C8yUrlSetting).into_diagnostic()?;
+    let url = build_proxy_url(host.as_str(), command.key())?;
     let jwt = Jwt::retrieve(&config)
         .await
         .context("Failed when requesting JWT from Cumulocity")?;
@@ -80,6 +77,12 @@ async fn proxy(command: RemoteAccessConnect, config: TEdgeConfig) -> miette::Res
     Ok(())
 }
 
+fn supported_operation_path(config_dir: &Path) -> PathBuf {
+    let mut path = config_dir.to_owned();
+    path.push("operations/c8y/c8y_RemoteAccessConnect");
+    path
+}
+
 fn build_proxy_url(cumulocity_host: &str, key: &str) -> miette::Result<Url> {
     format!("wss://{cumulocity_host}/service/remoteaccess/device/{key}")
         .parse()
@@ -87,67 +90,15 @@ fn build_proxy_url(cumulocity_host: &str, key: &str) -> miette::Result<Url> {
         .context("Creating websocket URL")
 }
 
-struct Websocket {
-    socket: WsStream<ConnectStream>,
-}
-
-fn generate_sec_websocket_key() -> String {
-    let mut rng = rand::thread_rng();
-    let mut bytes = [0u8; 16];
-    rng.fill_bytes(&mut bytes);
-    base64::encode(bytes)
-}
-
-impl Websocket {
-    async fn new(url: &Url, authorization: String) -> miette::Result<Self> {
-        let request = http::Request::builder()
-            .header("Authorization", authorization)
-            .header("Sec-WebSocket-Key", generate_sec_websocket_key())
-            .header("Host", url.host_str().unwrap())
-            .header("Connection", "Upgrade")
-            .header("Upgrade", "websocket")
-            .header("sec-websocket-version", "13")
-            .uri(url.to_string())
-            .body(())
-            .into_diagnostic()
-            .context("Instantiating Websocket connection")?;
-
-        let socket = async_tungstenite::tokio::connect_async(request)
-            .await
-            .into_diagnostic()
-            .context("Connecting to Websocket")?
-            .0;
-        Ok(Websocket {
-            socket: WsStream::new(socket),
-        })
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
 
     #[test]
-    fn generated_key_is_base64_encoded_16_byte_sequence() {
-        let key = generate_sec_websocket_key();
-
-        let decoded = base64::decode(key).unwrap();
-
-        assert_eq!(decoded.len(), 16);
-    }
-
-    #[test]
-    fn generated_key_is_ascii() {
-        let key = generate_sec_websocket_key();
-
-        assert!(key.is_ascii());
-    }
-
-    #[test]
-    fn generated_keys_are_unique_per_connection() {
-        let key_1 = generate_sec_websocket_key();
-        let key_2 = generate_sec_websocket_key();
-
-        assert_ne!(key_1, key_2);
+    fn default_supported_operation_path() {
+        assert_eq!(
+            supported_operation_path("/etc/tedge".as_ref()),
+            PathBuf::from("/etc/tedge/operations/c8y/c8y_RemoteAccessConnect")
+        );
     }
 }
